requirements.txt
# Core ML dependencies
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.17
torch-cluster>=1.6.1
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Data and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
networkx>=3.1
tqdm>=4.65.0

# Configuration and logging
pyyaml>=6.0
omegaconf>=2.3.0
wandb>=0.15.0
tensorboard>=2.13.0

# Evaluation and metrics
scikit-learn>=1.3.0
torchmetrics>=1.0.0

# Interactive demos
streamlit>=1.25.0
gradio>=3.40.0

# Development tools
black>=23.0.0
ruff>=0.0.280
pytest>=7.4.0
pre-commit>=3.3.0

# Optional: Advanced optimization
ortools>=9.6.0
cvxpy>=1.3.0
